Constraint.Checker

Description

Module to verify constraints on .t2 file.

Documentation

run :: [String] -> IO ()

run, entry point. Expect a file path for the output, the .t2 input file and a file containing constraints to check.

check

Arguments

:: NTree Cell
The tree.

-> [Cstr]
Constraints to check.

-> String
Results.

check, check all the constraints.

checkConstraint

Arguments

:: NTree Cell
The tree.

-> Cstr
The constraint.

-> Bool
True, if the tree respects the constraint, False otherwise.

checkConstraint, check if the constaint is consistent with the tree.

match

Arguments

:: Identificator
The identificator.

-> NTree Cell
The tree.

-> Bool
True, if the cell at the root of the tree matchs the identifcator. False, otherwise.

match, return True if the root cell matchs the identificator.

lookFor

Arguments

:: NTree Cell
The tree.

-> (NTree Cell -> Bool)
Function determining if we must keep a subtree.

-> [NTree Cell]
Subtrees matched by the function.

lookFor, return trees with a root node matched by the function.

notContainedId

Arguments

:: Identificators
Identificators.

-> NTree Cell
The tree.

-> Identificators
Identificators not matched.

notContainedId, return the list of identifcatos not matched in the tree.

notContainedId'

Arguments

:: NTree Cell
The tree.

-> Identificators
Identificators.

-> Identificators
Identificators not matched.

notContainedId', return the list of identifcatos not matched for the root node of the tree.

checkUnder

Arguments

:: Identificators
Identificators, matching items that must be in all subtrees.

-> [NTree Cell]
Subtrees.

-> Bool
True, if all the items matched by the identificators are in all the subtrees.

checkUnder, return True if all the items in the first list are in all subtrees selected.

checkUnder'

Arguments

:: Identificators
Identificators, matching items.

-> [NTree Cell]
Subtrees.

-> Bool
True, if all the items matched by the identificators are in a subtree or none of them. False, otherwise.

checkUnder', return True if all the items in the first list are in a subtree selected, or none of the items are in.

checkUnder1

Arguments

:: Identificators
Identificators, matching items.

-> [NTree Cell]
Subtrees.

-> Bool
True if at least one items in the first list is in the subtree (for each subtrees). False, otherwise.

checkUnder1, return True if at least one items in the first list is in the subtree (for each subtrees).

containsId

Arguments

:: Identificators
Identificators.

-> Identificators
Identificators remaining to check for the root node.

-> Int
Number of matchs wanted.

-> NTree Cell
The tree.

-> Bool
True, if there is a number of matched wanted. False, otherwise.

containsId, check if a given number of identificators are matched.

containsId'

Arguments

:: Identificators
Identificators.

-> Int
Number of match.

-> [NTree Cell]
Subtrees.

-> Bool

containsId', check if a given number of identificators are matched in at least one of the subtrees.

checkNotUnder'

Arguments

:: Identificators
Identificators, matching items.

-> [NTree Cell]
Subtrees.

-> Bool
True if only one items in the first list is in the subtree (for each subtrees). False, otherwise.

checkNotUnder', return True if only one item in the first list is in a subtree selected, or none of the items are in.

checkBefore

Arguments

:: Identificators
Identificators.

-> [NTree Cell]
Subtrees.

-> Bool

checkBefore, check if the first item is before the second item (with left right traversal on root nodes of trees).