Editor.Editor

Description

Module to generate .t2 file for the pattern Editor.

Documentation

type Editor = String

type Topic = String

type Section = String

data Info

Constructors

Info
Info, contain information needed for the pattern.

Fields

infoParticipants :: ParticipantObjects

Participants.

infoNbSec :: Int

Number of section per passage.

infoNbSecEditor :: Int

Number of editors per section.

run :: [String] -> IO ()

run, entry point. Expect a file path in argument pointing to a file containing information needed.

generateLevels

Arguments

:: Info
Pattern info.

-> [Level]
Levels.

generateLevels, generate levels of the tree.

generateLvlTopics

Arguments

:: [Participant]
Participants.

-> Level
Topic's level.

generateLvlTopics, generate the topic level.

generateLvlTopics'

Arguments

:: [Participant]
Participants.

-> Int
Numbring of the topic.

-> [[[Topic]]]
Topics.

generateLvlTopics', cf generateLvlTopics.

generateLvlEditor

Arguments

:: [Participant]
Participants.

-> Level
Editor's level.

generateLvlEditor, generate the editor's level.

generateLvlEditor'

Arguments

:: [Participant]
Participants.

-> [[[Editor]]]
Editor.

generateLvlEditor', cf generateLvlEditor.

generateLvlSection

Arguments

:: [Participant]
Participants.

-> Int
Number of section.

-> Level
Section's level.

generateLvlSection, generate the section's level.

generateLvlSection'

Arguments

:: [Participant]
Participants.

-> Int
Number of section.

-> Int
Numbering of the topic.

-> [[[Section]]]
Sections.

generateLvlSection', cf generateLvlSection.

generateLvlSection''

Arguments

:: Int
Number of sections.

-> Int
Numbering of the topic.

-> Int
Section's numbering.

-> [[Section]]
Sections.

generateLvlSection'', cf generateLvlSection.

generateLvlEditor2

Arguments

:: [Participant]
Participants.

-> Int
Number of editors per section.

-> Int
Number of sections.

-> Level
Second editor level.

generateLvlEditor2, generate level for the sub-editors.

generateLvlEditor2'

Arguments

:: Int
Number of editors per section.

-> Int
Number of editors per section.

-> [Participant]
Participants remaining = topics remaining.

-> [Participant]
Participants.

-> [[[Editor]]]
Sub-editors.

generateLvlEditor2', cf generateLvlEditor2.

takeParticipants

Arguments

:: [Participant]
Participants.

-> Participant
Participant unwanted.

-> Int
Number of participants wanted.

-> ([Participant], [Participant])
Participants taken, participants remaining.

takeParticipants, take the number of participants wanted from the list, without the participant given.