Jigsaw.Jigsaw

Description

Module to generate .t2 file for the pattern Jigsaw.

Documentation

run :: [String] -> IO ()

run, entry point. Expect a file path in argument pointing to a file containing information needed.

notPossible

Arguments

:: [Possible a]
Possible actions.

-> [Possible String]
Not possible actions.

notPossible, return the list of not possible actions.

generateLevels

Arguments

:: Info
Pattern information.

-> ([Level], PatternObjects)
Levels and pattern objects generated.

generateLevels, generatee the levels of the tree.

generateActivitiesLvl

Arguments

:: Level
Activities level.

generateActivitiesLvl, generate activities level.

generateGroupsLvl

Arguments

:: Info
Pattern information.

-> Level
Groups level.

generateGroupsLvl, generate groups level.

createGroupsInitial

Arguments

:: Info
Pattern information.

-> [[Group]]
Initial groups.

createGroupsInitial, create initial groups

createGroupsExpert

Arguments

:: [Theme]
Themes.

-> [[Group]]
Expert groups.

createGroupsExpert, create expert groups.

createGroupsJigsaw

Arguments

:: Info
Pattern information.

-> [[Group]]
Jigsaw groups.

createGroupsJigsaw, create jigsaw groups.

generateParticipantsLvl

Arguments

:: Info
Pattern information.

-> Level
Participants level.

generateParticipantsLvl, generate participants level.

createParticipantsInitial

Arguments

:: Info
Pattern information.

-> [InitialGroup]
Initial groups.

createParticipantsInitial, split participants into initial groups

createParticipantsInitialWithTheme

Arguments

:: Info
Pattern information.

-> [(Theme, InitialGroup)]
Initial groups with their associated theme.

createParticipantsInitialWithTheme, split participants into initial groups.

createParticipantsInitial'

Arguments

:: [ExpertGroup]
Expert groups.

-> [Theme]
Themes.

-> [[Int]]
Repartition, for each expert groups in initial groups.

-> [(Theme, InitialGroup)]
Initial groups with their associated theme.

createParticipantsInitial', split participants into initial groups from expert groups.

createParticipantsExpert

Arguments

:: [Participant]
Participants.

-> [Theme]
Themes.

-> [ExpertGroup]
Expert groups.

createParticipantsExpert, split participants into expert groups. Merge participants splited per themes for each jigsaw group, into a list of participant splited only per themes

createParticipantsExpert'

Arguments

:: [JigsawGroup]
Repartition for jigsaw groups.

-> [Theme]
Themes.

-> [[[Participant]]]
Participants splited per jigsaw groups and per themes.

createParticipantsExpert', split jigsaw groups per themes

createParticipantsExpert''

Arguments

:: [Theme]
Themes.

-> JigsawGroup
Participants of one jigsaw group.

-> [[Participant]]
Participants splited between themes.

createParticipantsExpert'', split a Jigsaw group between themes.

createParticipantsJigsaw

Arguments

:: [Participant]
Participants.

-> [Theme]
Themes.

-> [JigsawGroup]
Jigsaw groups.

createParticipantsJigsaw, split participants into groups jigsaw.

generateResourcesLvl

Arguments

:: Info
Pattern information.

-> Level
The level resources.

generateResourcesLvl, generate resources level.

createResourcesInitial

Arguments

:: Info
Pattern information.

-> [[Resource]]
Resources.

createResourcesInitial, create resources list for the expert level

createResourcesInitial'

Arguments

:: [Int]
Number of groups for each theme.

-> [[Resource]]
List of resources for each theme.

-> [[Resource]]
List of resources duplicated for each theme by the number of initial groups for the theme.

createResourcesInitial', create resources list for the initial level.

createResourcesExpert

Arguments

:: [Participant]
Participants.

-> [Theme]
Themes.

-> [[Resource]]
Resources.

createResourcesExpert, create resources list for the expert level

createResourcesExpert'

Arguments

:: [Theme]
Themes.

-> [ExpertGroup]
Expert groups.

-> [[Resource]]
Resources.

createResourcesExpert', create resources for the expert level.

createResourcesJigsaw

Arguments

:: [Participant]
Participants.

-> [Theme]
Themes.

-> [[Resource]]
Resources for the jigsaw level.

createResourcesJigsaw, create resources list for the jigsaw level

repartitionInitial

Arguments

:: Info
Pattern information.

-> [Possible [Int]]
Possible repartition.

repartitionInitial, repartition for initial groups.

repartitionInitial'

Arguments

:: [ExpertGroup]
Expert groups.

-> [Theme]
Themes.

-> Int
Number of participants per groups.

-> Int
Above margin.

-> Int
Below margin.

-> [Possible [Int]]
List of size for jigsaw's group.

repartitionInitial', list of size for jigsaw's group.

repartitionJigsaw

Arguments

:: [Participant]
Participants.

-> [Theme]
Themes.

-> Possible [Int]
Possible repartition.

repartitionJigsaw, list of size for jigsaw's group.

allResources

Arguments

:: [Theme]
Themes.

-> [String]
All resources.

allResources, create a list with all the resources.

nbParticipantJigsaw

Arguments

:: [Theme]
Themes.

-> (Int, Int, Int)
(number of participants wanted for a jigsaw group, how many participants additional are tolerated, how many participants less are tolerated)

nbParticipantJigsaw, give information on the repartition wanted for jigsaw groups.

nbGroupsInitial

Arguments

:: Info
Pattern info.

-> Int
Number of initial groups.

nbGroupInitial, number of initial groups.

nbGroupsInitialPerThemes

Arguments

:: Info
Pattern info.

-> [(Theme, [Int])]
(theme, length of groups of the theme)

nbGroupsInitialPerThemes, number of initial group per themes.

participantsList

Arguments

:: Info
Pattern information.

-> [Participant]
Participants.

participantsList, list of participants logins

groupObjectsList

Arguments

:: Info
Pattern information.

-> GroupObjects
Groups objects for .t2.

groupObjectsList, create the group objects list for .t2