JigsawModel

Description

Data strcuture used by Jigsaw.hs

Documentation

type InitialGroup = [Participant]

Initial Group

type ExpertGroup = [Participant]

Expert Group

type JigsawGroup = [Participant]

Jigsaw Group

data Theme

Constructors

Theme
Theme, a theme in jigsaw's context.

Fields

name :: String

Name of the theme.

resources :: [Resource]

Resources of the theme.

nbExpert :: Int

Number of expert for the theme needed in a Jigsaw group.

lowerMargin :: Int

Number of additionnal experts allowed in a Jigsaw group.

upperMargin :: Int

Number of less experts allowed in a Jigsaw group.

nbResources :: Int

Number of resources of this theme in an initial group.

data Info

Constructors

Info
Info, information on the pattern.

Fields

objects :: PatternObjects

Pattern objects.

themes :: [Theme]

Themes.

nbPPG :: Int

Number of participants per groups.

above :: Int

Above margin.

below :: Int

Below margin.