RT.RT

Description

Module to generate .t2 file for the pattern Reciprocal Teaching.

Documentation

data Info

Constructors

Info
Info, contain information needed for the pattern.

Fields

objects :: PatternObjects

Pattern's objects.

nbPPG :: Int

Number of participants per group preferred.

above :: Int

Above margin.

below :: Int

Below margin.

uniform :: Bool

Uniform repartition over closest one, if True.

run :: [String] -> IO ()

run, entry point. Expect a file path in argument pointing to a file containing information needed.

generateLevels

Arguments

:: Info
Information for the pattern.

-> ([Level], PatternObjects)

Level

= list of levels, PatternObjects = pattern objects.

generateLevels, generatee the levels for the tree.

generateActivityLvl

Arguments

:: Level
The activity level.

generateActivityLvl, generate the activity level.

generateResourceLvl

Arguments

:: [Resource]
List of resources corresponding to passages.

-> Level
The resource's level.

generateResourceLvl, generate the resource level.

generateGroupLvl

Arguments

:: Int
Number of passages (= number of resources).

-> [String]
List of groups for the learning activity.

-> Level
The group's level.

generateGroupLvl, generate the group level.

generateRoleLvl

Arguments

:: Int
Number of groups for the learning activity.

-> Level
The role's level.

generateRoleLvl, generate the role level.

generateParticipantLvl

Arguments

:: [Participant]
List of participants.

-> [Int]
Repartition of participants for the learning activity.

-> Int
Number of passages (= number of resources).

-> Level
Participant's level.

generateParticipantLvl, generate the participant level.

roundRobin

Arguments

:: Int
Number of time that it must turns the list (= number of passages = number of resources).

-> [[Participant]]
Lists of participants already splitted in groups.

-> [[[Participant]]]
The part of the participant's level for the learning activity.

roundRobin, create the part of the participant's level for the learning activity. It turns the list of participants for each group, and take the first participant apart to be the teacher.

turn :: [[Participant]] -> [[Participant]]

turn, put the last participant of each group in the first place.

createGroups

Arguments

:: Int
Number of groups.

-> [String]
List of groups.

createGroups, create groups for the learning level.

repartitionParticipant

Arguments

:: Int
Number of participants.

-> Int
Number of participants per group wanted.

-> Int
Above margin.

-> Int
Below margin.

-> Bool
If the user prefers a uniform repartition or a closest repartition.

-> Possible [Int]
Possible size's repartition.

repartitionParticipant, size repartition for groups of the learning level.