Simulation.Simulation

Description

Module to generate .t2 file for the pattern Simulation.

Documentation

run :: [String] -> IO ()

run, entry point. Expect a file path in argument pointing to a file containing information needed.

generateLevels

Arguments

:: Info
Information for the pattern.

-> ([Level], PatternObjects)

Level

= list of levels, PatternObjects = pattern objects.

generateLevels, generatee the levels for the tree.

generateActivityLvl

Arguments

:: Level
The activity level.

generateActivityLvl, generate the activity level.

generateGroupLvl

Arguments

:: [String]
Groups for the group simulation activity.

-> Level
The group's level.

generateGroupLvl, generate the group level.

generateRoleLvl

Arguments

:: Int
Number of groups for the simulation group activity.

-> [Role]
Roles' names.

-> Level
The role's level.

generateRoleLvl, generate the role level.

generateParticipantLvl

Arguments

:: [Participant]
List of participants.

-> [[Int]]
Repartition of participants for the simulation group activity.

-> Level
Participant's level.

generateParticipantLvl, generate the participant level.

generateResourceLvl

Arguments

:: [Resource]
List of resource for the problem.

-> [RoleSimu]
List of roles.

-> Int
Number of groups.

-> Int
Number of participants.

-> [[Int]]
Repartition by groups and roles for the simulation group activity.

-> Level
Participant's level.

generateResourceLvl, generate the resource level.

createResourcesForRoles

Arguments

:: [RoleSimu]
List of roles.

-> [Int]
Number of participants per role.

-> [[[Resource]]]
Resources for each students for his role.

createResourcesForRoles, create resources for each students for his role.

mergeRoles

Arguments

:: [[[Participant]]]
Participants divided by roles.

-> [[[Participant]]]
Participants divided by groups and roles.

-> Int
Number of roles.

-> [[[Participant]]]
Participants divided by roles

mergeRoles, merge participants of the same role from groups of simulation group activity.

createParticipantsRoles

Arguments

:: [[[Participant]]]
Participants divided in simulation groups.

-> [[Int]]
Splits by role for each simulation group.

-> [[[Participant]]]
Participants divided by role.

createParticipantsRoles, create the list of group of participants divided by roles.

createGroups

Arguments

:: Int
Number of groups.

-> [String]
List of groups.

createGroups, create groups for the simulation group activity.

repartitionGroups

Arguments

:: Int
The number of participants.

-> [RoleSimu]
Roles.

-> Possible [[Int]]
Repartition.

repartitionGroups, give the number of participant needed per role per group.