Util.KobbeComponents

Description

Module to have types and methods for Kobbe components.

Documentation

type Activity = String

Activity

type Group = String

Group

type Participant = String

Participant

type Role = String

Role

type Resource = String

Resource

type PatternObjects = (ActivityObjects, GroupObjects, ParticipantObjects, ResourceObjects, RoleObjects)

PatternObjects represents the list of objects for the bases notions

type ActivityObjects = [(Name, Description)]

ActivityObjects

type GroupObjects = [(Name, Description)]

GroupObjects

type ParticipantObjects = [(Login, Firstname, Surname, Email, City, Country)]

ParticipantObjects

type ResourceObjects = [(Name, Description, MoodleTableName, MoodleResourceID)]

ResourceObjects

type RoleObjects = [(Name, Description)]

RoleObjects

type TeacherNotes = [String]

TeacherNotes

type Name = String

type Description = String

type Login = String

type Firstname = String

type Surname = String

type Email = String

type Country = String

type City = String

type MoodleTableName = String

type MoodleResourceID = String

participantsLogins

Arguments

:: ParticipantObjects
Participants objects.

-> [Participant]
Participants logins.

participantsLogins, return the list of participants logins.

resourcesNames

Arguments

:: ResourceObjects
Resources objects.

-> [Resource]
Resources names.

resourcesNames, returns the list of resources names.

rolesNames

Arguments

:: RoleObjects
Roles objects.

-> [Role]
Roles names.

rolesNames, returns the list of roles names.

activitiesObjects :: PatternObjects -> ActivityObjects

activitiesObjects, return the activities objects from a pattern object.

groupsObjects :: PatternObjects -> GroupObjects

groupsObjects, return the groups objects from a pattern object.

participantsObjects :: PatternObjects -> ParticipantObjects

participantsObjects, return the participants objects from a pattern object.

resourcesObjects :: PatternObjects -> ResourceObjects

resourcesObjects, return the resources objects from a pattern object.

rolesObjects :: PatternObjects -> RoleObjects

rolesObjects, return the roles objects from a pattern object.

showObjects

Arguments

:: PatternObjects
Pattern objects.

-> String
Pattern objects in .t2 format.