Util.Tree

Description

Module containing the data structure for the tree and useful functions on tree.

Documentation

type CellLabel = String

Name of the column (notion), in which the cell is.

type CellNumbering = String

The number of the cell (differents cells can have the same number across the tree).

type CellFatherNumbering = String

The number of the father, null for the root node.

type CellComponents = [String]

The component identification array.

type Cell = (CellLabel, CellNumbering, CellFatherNumbering, CellComponents)

Cell represents a node

data NTree a

Constructors

Node a [NTree a]
Data structure representing a tree. 'Node a' for the node and '[NTree a]' for the subtrees

showTree

Arguments

:: NTree Cell
The tree to convert.

-> String
The tree in .t2 format.

showTree, convert a tree in a string in .t2 format.

node

Arguments

:: NTree Cell
The tree.

-> Cell
The root's cell of the tree.

node, return the cell of a 'NTree Cell'.

subtrees

Arguments

:: NTree a
The tree.

-> [NTree a]
Subtrees of the tree.

subtrees, return subtrees of a NTree.

cellComponents

Arguments

:: Cell
The cell.

-> CellComponents
components of the cell.

cellComponents, return the components of a cell.

nbLeaf

Arguments

:: Eq a

=> NTree a
The tree.

-> Int
Number of leaves for the tree.

nbLeaf, return the number of leaves for the given tree