Util.Util

Documentation

data Possible a

Constructors

Possible a
If the action if possible, with the result.

NotPossible String
If the action is impossible with a error message.

readInt

Arguments

:: String
The number.

-> Int
Int value.

readInt, read an int

addAll :: [[a]] -> [[a]] -> [[a]]

readText

Arguments

:: String
Text of the information file.

-> (String -> a)
Function to convert the second line in a data containing information for the pattern

-> (String, a)
(output file, pattern's information).

readText, read the information file.

createList

Arguments

:: Int
Number of participants to be created

-> String
List of participants.

createList, Create a list a participants for test.

writeT2

Arguments

:: String
File path.

-> [String]
List of notions names.

-> NTree Cell
The tree.

-> String
Pattern objects.

-> IO ()

writeT2, write the entire .t2 file.

writeT2Err

Arguments

:: String
File path.

-> [Possible a]
List of errors.

-> IO ()

writeT2Err, write the .t2 error file.

errors

Arguments

:: [Possible a]
Errors.

-> String
Errors in .t2.

errors, convert errors in a .t2 format.

errors'

Arguments

:: [Possible a]
Errors.

-> Bool
If it's the first error.

-> String
Errors in .t2.

readHeader

Arguments

:: String
.t2 content.

-> String
First line of the .t2.

readHeader, return the first line of .t2 file.

stripHeader

Arguments

:: String
.t2 content.

-> String
.t2 content without the head (notions names + script=).

stripHeader, strip the header of .t2 content.

stripFooter

Arguments

:: [String]
The file content in lines.

-> [String]
All lines without footer's lines.

readTree

Arguments

:: String
File path for .t2.

-> IO (NTree Cell)
The tree stored in the .t2.

readTree, read the tree from .t2.

readConstraints

Arguments

:: String
File path.

-> IO [Cstr]
Constraints.

readConstraints, read constraints stored in a file

splitList

Arguments

:: [a]
The list to split.

-> [Int]
List of groups sizes.

-> [[a]]
The list splitted.

splitList, split a list into group of size given by the list.

splitList2

Arguments

:: [a]
The list to split.

-> Int
Number of sublists.

-> Int
Size of the split.

-> [[a]]
The list splitted.

splitList2, split a list into group of given size.

splitList2' :: [a] -> Int -> Int -> [[a]]

possibleToList

Arguments

:: Possible [a]
A possible action.

-> [a]
The result of the action, [] if not possible.

possibleToList, convert 'Possible [a]' to '[a]'.

repartition

Arguments

:: Int
Number to split.

-> Int
Size of a split wanted.

-> Int
Above margin tolerated.

-> Int
Below margin tolerated.

-> Bool
True, if the user prefers splits uniform, no matter if there is a huge variation with the size wanted. False if he wants the solution the closest.

-> Possible [Int]
Possible size repartition.

repartition, give a list of group size, for the best repartition.

repartition'

Arguments

:: Int
Number to split.

-> Int
Size of a split wanted.

-> Int
Above margin tolerated.

-> Int
Below margin tolerated.

-> Int
Variation.

-> (Bool, [Int])
If a repartition is possible then it returns (True, repartition) else (False, []).

repartition', tries to create a repartion with less variations possible.

repartitionUniform

Arguments

:: Int
Number to split.

-> Int
Size of a split wanted.

-> Int
Above margin tolerated.

-> Int
Below margin tolerated.

-> Int
Variation.

-> (Bool, [Int])
If a repartition is possible then it returns (True, repartition) else (False, []).

repartitionUniform, repartition with the same size for each split, and with the closest possible size of the size wanted.

repartition2Multiple

Arguments

:: Int
Number to split.

-> [(Int, Int, Int)]
List of splits with for each split : the size wanted, the above margin and the below margin.

-> Possible [[Int]]
Possible size repartition, list of list of splits.

repartition2

Arguments

:: Int
Number to split.

-> [(Int, Int, Int)]
List of splits with for each split : the size wanted, the above margin and the below margin. Each split is done once.

-> Possible [Int]
Possible size repartition.

sumTriplets

Arguments

:: Num a

=> [(a, a, a)]
Triplets.

-> (a, a, a)
The sum.

sumTriplets, sum a list of triplet. sumTriplet [(1,2,4),(-1,0,5)] => (0,2,9).

sumList :: Num a => [a] -> [a] -> [a]

sumList, do the sum of 2 lists. Ex : sumList [1,2] [2,3,4] => [3,5,4]

repartition2'

Arguments

:: Int
.

-> Int
.

-> [(Int, Int, Int)]
.

-> [Int]
.

repartition2', cf repartition2, expect that repartition2' is used only when there is a solution and the second parameter take the difference between the number to split and the sum of sizes wanted