Util.Util

Documentation

type Activity = String

Activity

type Group = String

Group

type Participant = String

Participant

type Role = String

Role

type Resource = String

Resource

type PatternObjects = (ActivityObjects, GroupObjects, ParticipantObjects, ResourceObjects, RoleObjects)

PatternObjects represents the list of objects for the bases notions

type ActivityObjects = [(Name, Description)]

ActivityObjects

type GroupObjects = [(Name, Description)]

GroupObjects

type ParticipantObjects = [(Login, FirstName, Surname, Email, City, Country)]

ParticipantObjects

type ResourceObjects = [(Name, Description, MoodleTableName, MoodleResourceID)]

ResourceObjects

type RoleObjects = [(Name, Description)]

RoleObjects

type TeacherNotes = [String]

TeacherNotes

type Name = String

type Description = String

type Login = String

type FirstName = String

type Surname = String

type Email = String

type Country = String

type City = String

type MoodleTableName = String

type MoodleResourceID = String

data Possible a

Constructors

Possible a
If the action if possible, with the result.

NotPossible String
If the action is impossible with a error message.

readInt

Arguments

:: String
The number.

-> Int
Int value.

readInt, read an int

participantsLogins

Arguments

:: ParticipantObjects
Participants objects.

-> [Participant]
Participants logins.

participantsLogins, return the list of participants logins.

resourcesNames

Arguments

:: ResourceObjects
Resources objects.

-> [Resource]
Resources names.

resourcesNames, returns the list of resources names.

activitiesObjects :: PatternObjects -> ActivityObjects

activitiesObjects, return the activities objects from a pattern object.

groupsObjects :: PatternObjects -> GroupObjects

groupsObjects, return the groups objects from a pattern object.

participantsObjects :: PatternObjects -> ParticipantObjects

participantsObjects, return the participants objects from a pattern object.

resourcesObjects :: PatternObjects -> ResourceObjects

resourcesObjects, return the resources objects from a pattern object.

rolesObjects :: PatternObjects -> RoleObjects

rolesObjects, return the roles objects from a pattern object.

createList

Arguments

:: Int
Number of participants to be created

-> String
List of participants.

createList, Create a list a participants for test.

writeT2

Arguments

:: String
File path.

-> [String]
List of notions names.

-> NTree Cell
The tree.

-> PatternObjects
Pattern objects.

-> IO ()

writeT2, write the entire .t2 file.

writeT2Err

Arguments

:: String
File path.

-> [Possible a]
List of errors.

-> IO ()

writeT2Err, write the .t2 error file.

showObjects

Arguments

:: PatternObjects
Pattern objects.

-> String
Pattern objects in .t2 format.

errors

Arguments

:: [Possible a]
Errors.

-> String
Errors in .t2.

errors, convert errors in a .t2 format.

errors'

Arguments

:: [Possible a]
Errors.

-> Bool
If it's the first error.

-> String
Errors in .t2.

readHeader

Arguments

:: String
.t2 content.

-> String
First line of the .t2.

readHeader, return the first line of .t2 file.

stripHeader

Arguments

:: String
.t2 content.

-> String
.t2 content without the head (notions names + script=).

stripHeader, strip the header of .t2 content.

stripFooter

Arguments

:: [String]
The file content in lines.

-> [String]
All lines without footer's lines.

readTree

Arguments

:: String
File path for .t2.

-> IO (NTree Cell)
The tree stored in the .t2.

readTree, read the tree from .t2.

readObjects

Arguments

:: String
File path to the information file.

-> (String, PatternObjects)
(file path for output, pattern objects).

readObjects, read informations from .txt information file.

readConstraints

Arguments

:: String
File path.

-> IO [Cstr]
Constraints.

readConstraints, read constraints stored in a file

splitList

Arguments

:: [a]
The list to split.

-> [Int]
List of groups sizes.

-> [[a]]
The list splitted.

splitList, split a list into group of size given by the list.

splitList2

Arguments

:: [a]
The list to split.

-> Int
Number of sublists.

-> Int
Size of the split.

-> [[a]]
The list splitted.

splitList2, split a list into group of given size.

splitList2' :: [a] -> Int -> Int -> [[a]]

possibleToList

Arguments

:: Possible [a]
A possible action.

-> [a]
The result of the action, [] if not possible.

possibleToList, convert 'Possible [a]' to '[a]'.

repartition

Arguments

:: Int
Number to split.

-> Int
Size of a split wanted.

-> Int
Above margin tolerated.

-> Int
Below margin tolerated.

-> Bool
True, if the user prefers splits uniform, no matter if there is a huge variation with the size wanted. False if he wants the solution the closest.

-> Possible [Int]
Possible size repartition.

repartition, give a list of group size, for the best repartition.

repartition'

Arguments

:: Int
Number to split.

-> Int
Size of a split wanted.

-> Int
Above margin tolerated.

-> Int
Below margin tolerated.

-> Int
Variation.

-> (Bool, [Int])
If a repartition is possible then it returns (True, repartition) else (False, []).

repartition', tries to create a repartion with less variations possible.

repartitionUniform

Arguments

:: Int
Number of participants to split.

-> Int
Number of participants per groups wanted.

-> Int
Above margin tolerated.

-> Int
Below margin tolerated.

-> Int
Variation.

-> (Bool, [Int])
If a repartition is possible then it returns (True, repartition) else (False, []).

repartitionUniform, repartition with the same size for each group, and with the closest possible size of the size wanted.

repartition2

Arguments

:: Int
Number to split.

-> [(Int, Int, Int)]
List of splits with for each split : the size wanted, the above margin and the below margin.

-> Possible [Int]
Possible size repartition.

repartition2'

Arguments

:: Int
.

-> Int
.

-> [(Int, Int, Int)]
.

-> [Int]
.

repartition2', cf repartition2, expect that repartition2' is used only when there is a solution and the second parameter take the difference between the number to split and the sum of sizes wanted